Add fixes after latest changes in forked `rust-client`
Add fixes after latest changes in the forked `rust-client` repository
regarding a refinement in integer sizes.

+++ history/src/main.rs
@@ -28,7 +28,7 @@ fn main() {
     let args = Args::parse();
     let client = Client::new(args.rpc);
     let block = client.get_block_by_hash(&args.hash, false).unwrap();
-    if block.number != args.height as u64 {
+    if block.number != args.height {
         log::error!(
             height = args.height,
             hash = args.hash,
